+++ gems/aws-sdk-ec2/lib/aws-sdk-ec2/client.rb
@@ -27,8 +27,6 @@ require 'aws-sdk-core/plugins/client_metrics_plugin.rb'
 require 'aws-sdk-core/plugins/client_metrics_send_plugin.rb'
 require 'aws-sdk-core/plugins/transfer_encoding.rb'
 require 'aws-sdk-core/plugins/http_checksum.rb'
-require 'aws-sdk-core/plugins/dualstack_endpoint.rb'
-require 'aws-sdk-core/plugins/fips_endpoint.rb'
 require 'aws-sdk-core/plugins/signature_v4.rb'
 require 'aws-sdk-core/plugins/protocols/ec2.rb'
 require 'aws-sdk-ec2/plugins/copy_encrypted_snapshot.rb'
@@ -77,8 +75,6 @@ module Aws::EC2
     add_plugin(Aws::Plugins::ClientMetricsSendPlugin)
     add_plugin(Aws::Plugins::TransferEncoding)
     add_plugin(Aws::Plugins::HttpChecksum)
-    add_plugin(Aws::Plugins::DualstackEndpoint)
-    add_plugin(Aws::Plugins::FipsEndpoint)
     add_plugin(Aws::Plugins::SignatureV4)
     add_plugin(Aws::Plugins::Protocols::EC2)
     add_plugin(Aws::EC2::Plugins::CopyEncryptedSnapshot)
@@ -282,14 +278,6 @@ module Aws::EC2
     #
     #     ** Please note ** When response stubbing is enabled, no HTTP
     #     requests are made, and retries are disabled.
-    #
-    #   @option options [Boolean] :use_dualstack_endpoint
-    #     When set to `true`, dualstack enabled endpoints (with `.aws` TLD)
-    #     will be used if available.
-    #
-    #   @option options [Boolean] :use_fips_endpoint
-    #     When set to `true`, fips compatible endpoints
-    #     will be used if available.
     #
     #   @option options [Boolean] :validate_params (true)
     #     When `true`, request parameters are validated before