+++ gems/aws-sdk-kms/lib/aws-sdk-kms/client.rb
@@ -27,8 +27,6 @@ require 'aws-sdk-core/plugins/client_metrics_plugin.rb'
 require 'aws-sdk-core/plugins/client_metrics_send_plugin.rb'
 require 'aws-sdk-core/plugins/transfer_encoding.rb'
 require 'aws-sdk-core/plugins/http_checksum.rb'
-require 'aws-sdk-core/plugins/dualstack_endpoint.rb'
-require 'aws-sdk-core/plugins/fips_endpoint.rb'
 require 'aws-sdk-core/plugins/signature_v4.rb'
 require 'aws-sdk-core/plugins/protocols/json_rpc.rb'
 
@@ -75,8 +73,6 @@ module Aws::KMS
     add_plugin(Aws::Plugins::ClientMetricsSendPlugin)
     add_plugin(Aws::Plugins::TransferEncoding)
     add_plugin(Aws::Plugins::HttpChecksum)
-    add_plugin(Aws::Plugins::DualstackEndpoint)
-    add_plugin(Aws::Plugins::FipsEndpoint)
     add_plugin(Aws::Plugins::SignatureV4)
     add_plugin(Aws::Plugins::Protocols::JsonRpc)
 
@@ -288,14 +284,6 @@ module Aws::KMS
     #
     #     ** Please note ** When response stubbing is enabled, no HTTP
     #     requests are made, and retries are disabled.
-    #
-    #   @option options [Boolean] :use_dualstack_endpoint
-    #     When set to `true`, dualstack enabled endpoints (with `.aws` TLD)
-    #     will be used if available.
-    #
-    #   @option options [Boolean] :use_fips_endpoint
-    #     When set to `true`, fips compatible endpoints
-    #     will be used if available.
     #
     #   @option options [Boolean] :validate_params (true)
     #     When `true`, request parameters are validated before